9077: Move processOtherEntries into its own file, test cleanup and fixes
