+++ shared/src/business/useCases/elasticsearch/processOtherEntries.js
@@ -1,0 +1,8 @@
+const { processEntries } = require('./processEntries');
+
+export const processOtherEntries = ({ applicationContext, otherRecords }) =>
+  processEntries({
+    applicationContext,
+    recordType: 'otherRecords',
+    records: otherRecords,
+  });
